requirements.txt
Flask==2.0.3
waitress==2.1.1
certifi==2021.10.8
python-dotenv==0.19.2
Werkzeug~=2.0.3
dnspython==2.2.1
requests==2.27.1
beautifulsoup4==4.10.0
lxml~=4.8.0
git+https://github.com/p32929/mongoengine_mate-project.git
Flask-Cors==3.0.10
flask-jwt-extended==4.3.1
requests-toolbelt==0.9.1
Pillow==9.0.1
pytz==2022.1
Flask-APScheduler==1.12.3
Flask-Pydantic==0.9.0
pytest==7.1.2